refactor(core): extract pane layout logic to ChartPaneLayout
Move pane initialization, ratio management, resize boundary,
and pane CRUD operations to dedicated ChartPaneLayout class.
Chart delegates via layoutManager and keeps only signal fields
(_paneRatiosSignal, _paneLayoutSignal) which PaneLayout updates
through an onLayoutChange callback.

- chart.ts: -437 lines
- chartPaneLayout.ts: +461 lines (new)
